pyproject.toml
[project]
name = "rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.4",
    "fastapi>=0.121.1",
    "groq>=0.33.0",
    "inngest>=0.5.11",
    "llama-index>=0.14.7",
    "llama-index-core>=0.14.7",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-readers-file>=0.5.4",
    "ollama>=0.6.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "uvicorn>=0.38.0",
]
